fix: Coerce values to float

+++ src/the_n_body_problem/objects/body.py
@@ -3,9 +3,9 @@ import numpy as np
 
 class Body:
     def __init__(self, mass: float, pos: list[float], vel: list[float], colour: str):
-        self.mass = mass
-        self.pos = np.array(pos)
-        self.vel = np.array(vel)
+        self.mass = float(mass)
+        self.pos = np.array(pos).astype(float)
+        self.vel = np.array(vel).astype(float)
         self.colour = colour
         self.path = []
